requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
MetaTrader5==5.0.36
numpy==1.22.0
PyYAML==6.0
requests==2.27.1
schedule==1.1.0
scheduler==0.6.3
soupsieve==2.3.1
typeguard==2.13.3
urllib3==1.26.8
mysql-connector-python==8.0.27